remove dependency cycle between searchModule.js and localUtils.js
